+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to
 
 ## [Unreleased]
 
+## [2.3.0] - 2022-06-08
+
 ### Added
 
 - New [`inverse/2`][inverse-2] function
@@ -79,7 +81,8 @@ and this project adheres to
 
 Initial release
 
-[unreleased]: https://github.com/eproxus/mapz/compare/v2.2.0...HEAD
+[unreleased]: https://github.com/eproxus/mapz/compare/v2.3.0...HEAD
+[2.3.0]: https://github.com/eproxus/mapz/compare/v2.2.1...v2.3.0
 [2.2.0]: https://github.com/eproxus/mapz/compare/v2.1.1...v2.2.0
 [2.1.1]: https://github.com/eproxus/mapz/compare/v2.1.0...v2.1.1
 [2.1.0]: https://github.com/eproxus/mapz/compare/v2.0.0...v2.1.0